#29 test: 테스트 코드 수정
