+++ backend/src/test/java/codesquard/app/IntegrationTestSupport.java
@@ -1,5 +1,6 @@
 package codesquard.app;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -10,10 +11,10 @@ import codesquard.app.domain.category.CategoryRepository;
 import codesquard.app.domain.chat.ChatLogRepository;
 import codesquard.app.domain.chat.ChatRoomRepository;
 import codesquard.app.domain.image.ImageRepository;
-import codesquard.app.domain.interest.InterestRepository;
 import codesquard.app.domain.item.ItemRepository;
 import codesquard.app.domain.member.MemberRepository;
 import codesquard.app.domain.membertown.MemberTownRepository;
+import codesquard.app.domain.wish.WishRepository;
 
 @SpringBootTest
 public abstract class IntegrationTestSupport {
@@ -43,11 +44,23 @@ public abstract class IntegrationTestSupport {
 	protected ImageRepository imageRepository;
 
 	@Autowired
-	protected InterestRepository interestRepository;
+	protected WishRepository wishRepository;
 
 	@Autowired
 	protected ChatLogRepository chatLogRepository;
 
 	@Autowired
 	protected ChatRoomRepository chatRoomRepository;
+
+	@BeforeEach
+	void cleanup() {
+		chatLogRepository.deleteAllInBatch();
+		chatRoomRepository.deleteAllInBatch();
+		wishRepository.deleteAllInBatch();
+		imageRepository.deleteAllInBatch();
+		itemRepository.deleteAllInBatch();
+		categoryRepository.deleteAllInBatch();
+		memberRepository.deleteAllInBatch();
+		memberTownRepository.deleteAllInBatch();
+	}
 }
