+++ backend/src/test/java/codesquard/app/api/item/ItemFixedFactory.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 import codesquard.app.domain.category.Category;
 import codesquard.app.domain.image.Image;
-import codesquard.app.domain.interest.Interest;
 import codesquard.app.domain.item.Item;
 import codesquard.app.domain.item.ItemStatus;
 import codesquard.app.domain.member.Member;
+import codesquard.app.domain.wish.Wish;
 
 public class ItemFixedFactory {
 
@@ -19,13 +19,13 @@ public class ItemFixedFactory {
 	private static final String REGION = "가락 1동";
 	private static final LocalDateTime CREATED_AT = LocalDateTime.of(2023, 1, 1, 0, 0);
 
-	public static Item createFixedItem(Member member, Category category, List<Image> images, List<Interest> interests,
+	public static Item createFixedItem(Member member, Category category, List<Image> images, List<Wish> wishes,
 		Long viewCount) {
 		Item item = Item.create(TITLE, CONTENT, PRICE, STATUS, REGION, CREATED_AT, viewCount);
 		item.setMember(member);
 		item.setCategory(category);
 		images.forEach(item::addImage);
-		interests.forEach(item::addInterest);
+		wishes.forEach(item::addWish);
 		return item;
 	}
 }
