+++ backend/src/test/java/codesquard/app/api/item/ItemServiceTest.java
@@ -10,24 +10,21 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import codesquard.app.IntegrationTestSupport;
 import codesquard.app.api.image.ImageUploader;
 import codesquard.app.api.response.ItemListResponse;
 import codesquard.app.api.response.ItemResponse;
 import codesquard.app.domain.item.Item;
 import codesquard.app.domain.member.Member;
-import codesquard.support.DatabaseInitializer;
 import codesquard.support.SupportRepository;
 
 @SpringBootTest
@@ -39,15 +36,8 @@ class ItemServiceTest {
 	private EntityManager em;
 	@Autowired
 	private SupportRepository supportRepository;
-	@Autowired
-	private DatabaseInitializer databaseInitializer;
 	@MockBean
 	private ImageUploader imageUploader;
-
-	@AfterEach
-	public void tearDown() {
-		databaseInitializer.truncateTables();
-	}
 
 	@Test
 	@DisplayName("새로운 상품 등록에 성공한다.")
