+++ backend/src/test/java/codesquard/app/api/member/unit/MemberServiceTest.java
@@ -7,22 +7,18 @@ import java.nio.charset.StandardCharsets;
 
 import javax.persistence.EntityManager;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.transaction.annotation.Transactional;
 
-import codesquard.app.IntegrationTestSupport;
 import codesquard.app.api.image.ImageUploader;
 import codesquard.app.api.member.MemberService;
-import codesquard.app.api.oauth.OauthFixedFactory;
 import codesquard.app.domain.member.Member;
-import codesquard.support.DatabaseInitializer;
 
 @SpringBootTest
 class MemberServiceTest {
@@ -31,15 +27,8 @@ class MemberServiceTest {
 	private MemberService memberService;
 	@Autowired
 	private EntityManager em;
-	@Autowired
-	private DatabaseInitializer databaseInitializer;
 	@MockBean
 	private ImageUploader imageUploader;
-
-	@AfterEach
-	public void tearDown() {
-		databaseInitializer.truncateTables();
-	}
 
 	@Test
 	@DisplayName("프로필 사진 변경에 성공한다.")
