+++ backend/src/test/java/codesquard/app/api/wishitem/WishItemServiceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.*;
 
 import javax.persistence.EntityManager;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +12,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import codesquard.app.api.item.ItemRegisterRequest;
 import codesquard.app.domain.item.Item;
 import codesquard.app.domain.member.Member;
-import codesquard.support.DatabaseInitializer;
 import codesquard.support.SupportRepository;
 
 @SpringBootTest
@@ -25,13 +23,6 @@ class WishItemServiceTest {
 	private EntityManager em;
 	@Autowired
 	private SupportRepository supportRepository;
-	@Autowired
-	private DatabaseInitializer databaseInitializer;
-
-	@AfterEach
-	public void tearDown() {
-		databaseInitializer.truncateTables();
-	}
 
 	@Test
 	@DisplayName("관심상품 등록에 성공한다.")
